src/python/pyproject.toml
[tool.poetry]
name = "animoid-row"
version = "0.1.0"
description = ""
authors = ["aifa <antonis@typesystem.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.0"
langchain-openai = "^0.1.7"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
